+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Disallow incorrect comparison of atomic values in a non-atomic way.
 
+### Removed
+- Remove dependency on `golang.org/x/{lint, tools}`.
+
 ## [1.6.0] - 2020-02-24
 ### Changed
 - Drop library dependency on `golang.org/x/{lint, tools}`.