+++ CHANGELOG.md
@@ -63,7 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/uber-go/atomic/compare/v1.6.0...v1.7.0
+[1.7.0]: https://github.com/uber-go/atomic/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/uber-go/atomic/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/uber-go/atomic/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/uber-go/atomic/compare/v1.4.0...v1.5.0